Updated tense of bulleted list of git commands
I updated the bulleted list of git commands to use a more active voice and present tense. Previously, the "git push" bullet used past tense, which was inconsistent with the other bullets. We want to use active voice when possible, as per the HashiCorp style guide. I also removed the word "will" from the "git pull" bullet so that the definition would more closely resemble the phrasing of the other two bullets.