Updated synchronizing sentence to be more active
The sentence "When this happens, your local branch needs to be synchronized with the remote branch with git pull or git fetch and git merge." is a little bit passive. I updated the language to "When this happens, you need to synchronize your local branch with the remote branch using `git pull` or `git fetch` and `git merge`.", which is more active. I also added the back ticks (`) to the command terms so that they are consistently tagged throughout the page. I also added a space at the end of the sentence, since there was not one between the end of this sentence and the next one.