Updated wording in last sentence
Previously read "This is often what we desire to do, but some people prefer to use git fetch followed by git merge to make sure they understand the changes they are merging into their branch before the merge happens." I changed the word "desire" to "want" to avoid any possible confusion if this documentation is translated, and also just to use a more simple and direct word. I also added the back ticks (`) to the git commands in this sentence for consistency.