Updated definition for `git pull`
Upon second review of the definition of `git pull`, I decided to update the language of the first part of the definition to match that of `git fetch`, since a `git fetch` is the first part of a `git pull`. In practical terms, this meant changing the phrase "Get changes from a remote branch" to "Get changes from a remote repo".