Added a line break
I added a line break after the sentence When this happens, you need to synchronize your local branch with the remote branch using `git pull`, or `git fetch` and `git merge`. ". This felt like a natural place to add a break, which I think helps the overall readability of this content.